added prebuild to webapp, split precompute (culprit) from build
