+++ circuit-lib/circuit-lib.js/src/elgamal-babyjubjub/index.ts
@@ -1,2 +1,5 @@
 export * from "./elgamal";
 export * from "./pointEncoding";
+/// we're not default exporting precompute because it's a node only process.
+/// that's fine because precompute is only consumed in elgamal.test.ts
+/// we may change this when migrating to a bundler for zk.js
