+++ web/webapp/.gitignore
@@ -41,4 +41,4 @@ yarn-error.log*
 next-env.d.ts
 
 # dupe circuits
-# **/public/build-circuits/
+**/public/build-circuits/
